switching from deprecated ui-select2 -> ui-select
Implementing the cluster labels using multi-select ran into issues syncing between client/server,
as there's no mechanism within ui-select2 to trigger off of item removal. This would mean having to
constantly poll the server for the cluster labels on each change to make sure the client labels are properly
saved/deleted.

ui-select2 is also deprecated, and the project recommends switching to ui-select:
https://github.com/angular-ui/ui-select

Unfortunately, ui-select2 has a bug with angular versions <= 1.2.17, so we are upgrading to 1.2.18